Added books of abstracts for ATiP
I would like to have the pdfs for the books of abstracts for the Applied Topology in Poznan conference and workshop on the Dioscuri Website. This way we have a permanent link, even if we update the files, so we can have a QR code to the books of abstracts.